requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
alibabacloud-dysmsapi20170525==3.1.0
alibabacloud-tea==0.4.0
alibabacloud_credentials==0.3.6
alibabacloud_endpoint_util==0.0.3
alibabacloud_gateway_spi==0.0.2
alibabacloud_openapi_util==0.2.2
alibabacloud_tea_openapi==0.3.12
alibabacloud_tea_util==0.3.13
alibabacloud_tea_xml==0.0.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-dysmsapi==2.1.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
backports.zoneinfo==0.2.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
Django==4.2.17
django-cors-headers==4.4.0
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.5.0
idna==3.10
jmespath==0.10.0
multidict==6.1.0
pillow==10.4.0
propcache==0.2.0
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.2.3
xmltodict==0.14.2
yarl==1.15.2
# 富文本编辑器 5
django-ckeditor-5==0.2.11
# 后台美化
django-simpleui==2024.3.8
# 消息队列
celery>=5.3.6
redis>=5.0.1
django-redis>=5.4.0

# 支付相关
python-alipay-sdk>=3.3.0
wechatpy>=1.8.18
cryptography>=42.0.5
pycryptodome>=3.20.0

# ... 其他依赖 ...
openai>=1.0.0
httpx>=0.24.0
